requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==2.0.12
Django==5.0.4
django-auto-logout==0.5.1
django-cors-headers==4.3.1
django-filter==24.2
django-guardian==2.4.0
django-jazzmin==2.0.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
idna==3.7
Markdown==3.5.2
pillow==10.2.0
Pygments==2.4.0
PyJWT==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==5.1
requests==2.27.1
six==1.16.0
sqlparse==0.5.0
tzdata==2023.4
tzlocal==5.2
uritemplate==3.0.0
urllib3==1.26.18
